+++ src/engine/currencyEngine.js
@@ -28,9 +28,9 @@ import { calcFeesFromEarnCom, calcMinerFeePerByte } from './miningFees.js'
 import { broadcastFactories } from './broadcastApi.js'
 import { bns } from 'biggystring'
 import { getAllAddresses } from '../utils/formatSelector.js'
+import { InfoServer } from '../info/constants'
 import {
   addressToScriptHash,
-  keysFromWalletInfo,
   verifyTxAmount,
   sumUtxos
 } from '../utils/coinUtils.js'
@@ -63,8 +63,7 @@ export type EngineCurrencyInfo = {
 
   // Optional Settings
   forks?: Array<string>,
-  feeInfoServer?: string,
-  infoServer?: string
+  feeInfoServer?: string
 }
 
 export type CurrencyEngineSettings = {
@@ -154,21 +153,22 @@ export class CurrencyEngine {
     }
 
     const cachedRawKeys = await this.engineState.loadKeys()
-    const { seed, bip, coinType, rawKeys } = keysFromWalletInfo(
-      this.network,
-      this.walletInfo,
-      cachedRawKeys
-    )
+    const { master = {}, ...otherKeys } = cachedRawKeys || {}
+    const keys = this.walletInfo.keys || {}
+    const { format, coinType = -1 } = keys
+    const seed = keys[`${this.network}Key`]
+    const xpub = keys[`${this.network}Xpub`]
+    const rawKeys = { ...otherKeys, master: { xpub, ...master } }
 
     console.log(
-      `${this.walletId} - Created Wallet Type ${bip} for Currency Plugin ${
+      `${this.walletId} - Created Wallet Type ${format} for Currency Plugin ${
         this.pluginState.pluginName
       }`
     )
 
     this.keyManager = new KeyManager({
       seed: seed,
-      bip: bip,
+      bip: format,
       coinType: coinType,
       rawKeys: rawKeys,
       callbacks: callbacks,
@@ -282,14 +282,10 @@ export class CurrencyEngine {
   }
 
   async updateFeeTable () {
-    const { infoServer } = this.engineInfo
     try {
       await this.fetchFee()
-      if (!infoServer) {
-        throw new Error('infoServer not set')
-      }
       if (Date.now() - this.fees.timestamp > this.feeUpdateInterval) {
-        const url = `${infoServer}/networkFees/${this.currencyCode}`
+        const url = `${InfoServer}/networkFees/${this.currencyCode}`
         const feesResponse = await this.io.fetch(url)
         const feesJson = await feesResponse.json()
         if (validateObject(feesJson, InfoServerFeesSchema)) {