+++ src/info/constants.js
@@ -1,3 +1,4 @@
 // @flow
 
 export const imageServerUrl = 'https://developer.airbitz.co/content'
+export const InfoServer = 'https://info1.edgesecure.co:8444/v1'