+++ src/utils/jsonSchemas.js
@@ -65,7 +65,7 @@ export const electrumHeaderSchema = {
     merkle_root: { type: 'string' },
     timestamp: { type: 'number' },
     bits: { type: 'number' },
-    nonce: { type: 'number' }
+    nonce: { type: ['number', 'string'] }
   },
   required: ['block_height', 'timestamp']
 }